+++ modules/canvaskit/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - `SkPath.offset()`, `SkPath.drawOval`
  - `SkRRect` support (`SkCanvas.drawRRect`, `SkCanvas.drawDRRect`, `CanvasKit.RRectXY`).
    Advanced users can specify the 8 individual radii, if needed.
+ - `CanvasKit.computeTonalColors()`, which returns an SkColor.
 
 ### Changed
  - `MakeSkVertices` uses a builder to save a copy.